pyproject.toml
[tool.poetry]
name = "hw-monitor"
version = "0.1.0"
description = ""
authors = ["Max Gonzih <gonzih@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.10.2"
pyserial = "^3.5"
GPUtil = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
